backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
celery==5.4.0
redis==5.0.4
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.7.0
